Move bill-related elements to the 'bill' element group folder
