--- angular/components/elements/bill-input-fields/bill-input-fields.component.scss
@@ -1,10 +1,0 @@
-// bill-input-fields.component.scss
-@import "../../../styles/global-variables";
-@import "../../../styles/global-mixins";
-@import "../../../styles/global-functions";
-@import "../../../styles/global-placeholders";
-
-bill-input-fields {
-    @include flex(center, center, column);
-    width: 100%;
-}
