+++ angular/components/elements/element-group-bill/bill-input-field/bill-input-field.component.scss
@@ -1,8 +1,8 @@
 // bill-input-field.component.scss
-@import "../../../styles/global-variables";
-@import "../../../styles/global-mixins";
-@import "../../../styles/global-functions";
-@import "../../../styles/global-placeholders";
+@import "../../../../styles/global-variables";
+@import "../../../../styles/global-mixins";
+@import "../../../../styles/global-functions";
+@import "../../../../styles/global-placeholders";
 
 bill-input-field {
     @include flex(center, center, row);
