+++ angular/components/elements/element-group-bill/bill-input-field/bill-input-field.component.ts
@@ -1,8 +1,8 @@
 // bill-input-field.component.ts 
 import {Component, EventEmitter, Input, OnInit, Output} from "@angular/core";
-import {TagType} from "../../../models/html/TagType";
+import {TagType} from "../../../../models/html/TagType";
 import {BillInputFieldType} from "./models/BillInputFieldType";
-import {InputType} from "../ss-input/models/InputType";
+import {InputType} from "../../ss-input/models/InputType";
 import {BillInputContent} from "./models/BillInputContent";
 
 @Component({
