+++ angular/components/elements/element-group-bill/bill-input-field/models/BillInputContent.ts
