+++ angular/components/elements/element-group-bill/bill-input/bill-input.component.scss
@@ -1,8 +1,8 @@
 // bill-input.component.scss
-@import "../../../styles/global-variables";
-@import "../../../styles/global-mixins";
-@import "../../../styles/global-functions";
-@import "../../../styles/global-placeholders";
+@import "../../../../styles/global-variables";
+@import "../../../../styles/global-mixins";
+@import "../../../../styles/global-functions";
+@import "../../../../styles/global-placeholders";
 
 bill-input {
     position: absolute;
