+++ angular/components/elements/element-group-bill/bill-list/bill-list.component.scss
@@ -1,8 +1,8 @@
 // bill-list.component.scss
-@import "../../../styles/global-variables";
-@import "../../../styles/global-mixins";
-@import "../../../styles/global-functions";
-@import "../../../styles/global-placeholders";
+@import "../../../../styles/global-variables";
+@import "../../../../styles/global-mixins";
+@import "../../../../styles/global-functions";
+@import "../../../../styles/global-placeholders";
 
 bill-list {
     @include flex(flex-start, center, column);
