+++ angular/components/elements/element-group-bill/bill/models/types.ts
@@ -1,4 +1,4 @@
 import {EventEmitter} from "@angular/core";
-import {OneTimeBill, RecurringBill} from "../../../../models/paycheck/types";
+import {OneTimeBill, RecurringBill} from "../../../../../models/paycheck/types";
 
 export type BillEmitter = EventEmitter<OneTimeBill | RecurringBill>;
