+++ angular/components/elements/element-group-bill/recurring-bill-dropdown/models/RecurringBillTimeInterval.ts
