+++ angular/components/elements/income-input-field/income-input-field.component.scss
@@ -3,7 +3,7 @@
 @import "../../../styles/global-mixins";
 @import "../../../styles/global-functions";
 @import "../../../styles/global-placeholders";
-@import "../bill-input-field/bill-input-field.component";
+@import "../element-group-bill/bill-input-field/bill-input-field.component";
 
 income-input-field {
     @extend bill-input-field;
