--- angular/components/elements/recurring-bill-dropdown/recurring-bill-dropdown.component.scss
@@ -1,11 +1,0 @@
-// recurring-bill-dropdown.component.scss
-@import "../../../styles/global-variables";
-@import "../../../styles/global-mixins";
-@import "../../../styles/global-functions";
-@import "../../../styles/global-placeholders";
-
-recurring-bill-dropdown {
-    @include flex(center, center, row);
-    width: calc($fourteen-percent * 0.935);
-    height: $three-percent;
-}
